fix daos

+++ src/main/java/com/example/clientservererecepta/Client/StageHandler.java
@@ -96,7 +96,7 @@ public class StageHandler {
 
     private boolean isNumeric(String str) {
         try {
-            Integer.parseInt(str);
+            Double.parseDouble(str);
             return true;
         } catch (NumberFormatException e) {
             return false;
